Reproduce Bitcoin Core's handling of cookie files (#2)
Bitcoin Core uses a single call to std::getline to read the contents of
cookie files, making it ignore newlines in the file, as well as ignore
any lines after the first. This reproduces that behavior, so that all
cookie files that work with Bitcoin Core should work with this library.